\name{TPK-package}
\alias{TPK-package}
\alias{TPK}
\docType{package}
\title{
Utility functions for Tai Po Kau Forestry Plot 
}
\description{
Functions for Tai Po Kau Plot,eg. Data quality checking, Double Entry Validation. This package is still at early stage of development.
}
\details{
\tabular{ll}{
Package: \tab TPK\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-06-19\cr
License: \tab What license is it under?\cr
}
R Functions for Tai Po Kau forestry plot, currently only one function decheck() is available. 
}
\author{
Jinlong Zhang

Maintainer: Jinlong Zhang <jinlongzhang01@gmail.com>
}

\keyword{ package }

\examples{
## Coming soon. 
}
